Adding some refactorings
