+++ .travis.yml
@@ -2,7 +2,7 @@ dist: trusty
 sudo: required
 language: node_js
 node_js:
-  - "6.11.3"
+  - "8.11.1"
 
 install:
   - npm install
@@ -12,7 +12,7 @@ cache:
     - "node_modules"
 
 script:
-  - "npm test"
+  - "npm run test-on-travis"
 
 addons:
   sonarcloud:
