+++ package.json
@@ -1,13 +1,15 @@
 {
   "name": "sgt-fields",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "description": "Find the property in your object with Sargent Fields",
   "main": "sgt-fields.js",
   "scripts": {
     "mocha": "_mocha --",
-    "istanbul": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- -R spec && cat coverage/lcov.info | node node_modules/coveralls/bin/coveralls.js",
+    "istanbul": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- -R spec",
+    "istanbul-on-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- -R spec && cat coverage/lcov.info | node node_modules/coveralls/bin/coveralls.js",
     "coverage": "istanbul cover --print both node_modules/mocha/bin/_mocha --",
     "test": "set NODE_ENV=test && npm run istanbul",
+    "test-on-travis": "set NODE_ENV=test && npm run istanbul-on-travis",
     "posttest": "istanbul check-coverage --statements 100 --branches 100 --functions 100 --lines 100"
   },
   "repository": {
@@ -31,9 +33,9 @@
   "homepage": "https://github.com/rodrigocmoreira/sgt-fields#readme",
   "devDependencies": {
     "chai": "^4.1.2",
-    "coveralls": "^2.13.1",
+    "coveralls": "^3.0.1",
     "istanbul": "^0.4.5",
-    "mocha": "^3.5.1",
+    "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0"
   }
 }
